sparse: template on variable array type, make header only
