+++ src/nmodl/kinetic.cpp
@@ -857,10 +857,9 @@ for(_i=%d;_i<%d;_i++){\n",
                     static int first = 1;
                     if (first) {
                         first = 0;
-                        Sprintf(buf, "extern double *_getelm(int, int);\n");
-                        qv = linsertstr(procfunc, buf);
-                        Sprintf(buf, "extern double *_nrn_thread_getelm(SparseObj*, int, int);\n");
-                        vectorize_substitute(qv, buf);
+                        if (vectorize) {
+                            linsertstr(procfunc,"extern double *_nrn_thread_getelm(SparseObj*, int, int);\n");
+                        }
                     }
                 }
             }
